[LLVM][NVVM] Simplify IntrinsicDiagnosticsProvider
Replace the virtual base class + registration mechanism with a simple
static utility class whose query methods directly format the canonical
type information for any non-overloaded intrinsic.
